# template-api-nestjs/openapi/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]},"get":{"operationId":"UsersController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/v1/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/v1/files/upload":{"post":{"operationId":"FilesController_uploadFile","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["Files"],"security":[{"bearer":[]}]}},"/api/v1/files/{path}":{"get":{"operationId":"FilesController_download","parameters":[],"responses":{"200":{"description":""}},"tags":["Files"]}},"/api/v1/auth/email/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthEmailLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/admin/email/login":{"post":{"operationId":"AuthController_adminLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthEmailLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/email/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterLoginDto"}}}},"responses":{"204":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/email/confirm":{"post":{"operationId":"AuthController_confirmEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthConfirmEmailDto"}}}},"responses":{"204":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/forgot/password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthForgotPasswordDto"}}}},"responses":{"204":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/reset/password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResetPasswordDto"}}}},"responses":{"204":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]},"patch":{"operationId":"AuthController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]},"delete":{"operationId":"AuthController_delete","parameters":[],"responses":{"204":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"204":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/v1/auth/facebook/login":{"post":{"operationId":"AuthFacebookController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthFacebookLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/google/login":{"post":{"operationId":"AuthGoogleController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthGoogleLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/twitter/login":{"post":{"operationId":"AuthTwitterController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTwitterLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/apple/login":{"post":{"operationId":"AuthAppleController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthAppleLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/":{"get":{"operationId":"HomeController_appInfo","parameters":[],"responses":{"200":{"description":""}},"tags":["Home"]}}},"info":{"title":"API","description":"API docs","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"FileEntity":{"type":"object","properties":{"id":{"type":"string","example":"cbcfa8b8-3a25-4adb-a9c6-e325f0d0f3ae"}},"required":["id"]},"Role":{"type":"object","properties":{"id":{"type":"number","example":1},"name":{"type":"string","example":"Admin"}},"required":["id","name"]},"Status":{"type":"object","properties":{"id":{"type":"number","example":1},"name":{"type":"string","example":"Active"}},"required":["id","name"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"object","example":"test1@example.com"},"password":{"type":"string"},"firstName":{"type":"object","example":"John"},"lastName":{"type":"object","example":"Doe"},"photo":{"$ref":"#/components/schemas/FileEntity"},"role":{"$ref":"#/components/schemas/Role"},"status":{"$ref":"#/components/schemas/Status"}},"required":["email","password","firstName","lastName","photo","role","status"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"object","example":"test1@example.com"},"password":{"type":"string"},"firstName":{"type":"object","example":"John"},"lastName":{"type":"object","example":"Doe"},"photo":{"$ref":"#/components/schemas/FileEntity"},"role":{"$ref":"#/components/schemas/Role"},"status":{"$ref":"#/components/schemas/Status"}}},"AuthEmailLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"test1@example.com"},"password":{"type":"string"}},"required":["email","password"]},"AuthRegisterLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"test1@example.com"},"password":{"type":"string"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"}},"required":["email","password","firstName","lastName"]},"AuthConfirmEmailDto":{"type":"object","properties":{"hash":{"type":"string"}},"required":["hash"]},"AuthForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"AuthResetPasswordDto":{"type":"object","properties":{"password":{"type":"string"},"hash":{"type":"string"}},"required":["password","hash"]},"AuthUpdateDto":{"type":"object","properties":{"photo":{"$ref":"#/components/schemas/FileEntity"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"},"password":{"type":"string"},"oldPassword":{"type":"string"}},"required":["photo","firstName","lastName","password","oldPassword"]},"AuthFacebookLoginDto":{"type":"object","properties":{"accessToken":{"type":"string","example":"abc"}},"required":["accessToken"]},"AuthGoogleLoginDto":{"type":"object","properties":{"idToken":{"type":"string","example":"abc"}},"required":["idToken"]},"AuthTwitterLoginDto":{"type":"object","properties":{"accessTokenKey":{"type":"string","example":"abc"},"accessTokenSecret":{"type":"string","example":"abc"}},"required":["accessTokenKey","accessTokenSecret"]},"AuthAppleLoginDto":{"type":"object","properties":{"idToken":{"type":"string","example":"abc"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["idToken"]}}}}